+++ base/statistics.jl
@@ -224,7 +224,7 @@ varm(iterable, m::Number; corrected::Bool=true) =
 
 ## variances over ranges
 
-function varm(v::Range, m::Number)
+function varm(v::AbstractRange, m::Number)
     f  = first(v) - m
     s  = step(v)
     l  = length(v)
@@ -235,7 +235,7 @@ function varm(v::Range, m::Number)
     return vv
 end
 
-function var(v::Range)
+function var(v::AbstractRange)
     s  = step(v)
     l  = length(v)
     vv = abs2(s) * (l + 1) * l / 12
@@ -561,7 +561,7 @@ julia> middle(1:10)
 5.5
 ```
 """
-middle(a::Range) = middle(a[1], a[end])
+middle(a::AbstractRange) = middle(a[1], a[end])
 
 """
     middle(a)